Refactor into individual files.
